requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11.15
django-allauth==0.33.0
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-taggit==0.23.0
django-taggit-serializer==0.1.7
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.7
oauthlib==2.1.0
Pillow==5.3.0
pkg-resources==0.0.0
psycopg2-binary==2.8.3
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.24.1
whitenoise==4.1.2
